+++ src/tests/file_tests.rs
@@ -6,12 +6,18 @@ use super::*;
 fn test_set_log_path() {
     let logger = Logger::new();
     // Setting a path that doesn't exist should not work
-    assert!(!logger.set_log_path("this/path/is/not/real/file.log"));
+    assert_eq!(
+        logger.set_log_path("this/path/is/not/real/file.log"),
+        Err(SetLogPathError::CouldNotCreateLogFile)
+    );
     assert!(logger.get_log_path().is_none());
     assert!(!logger.has_log_writer());
 
     // Setting a directory that exists should not work
-    assert!(!logger.set_log_path("test_log_directory/"));
+    assert_eq!(
+        logger.set_log_path("test_log_directory/"),
+        Err(SetLogPathError::PathIsNotAFile)
+    );
     assert!(logger.get_log_path().is_none());
 
     // Ensure the other.log is not actually there still
@@ -24,7 +30,7 @@ fn test_set_log_path() {
     }
 
     // Directories exist, but file doesn't, should create the file
-    assert!(logger.set_log_path("test_log_directory/other.log"));
+    assert!(logger.set_log_path("test_log_directory/other.log").is_ok());
     let other_log_path = logger
         .get_log_path()
         .expect("logger log path should be set to other.log");
@@ -35,7 +41,7 @@ fn test_set_log_path() {
     logger.info("Log file created: other.log");
 
     // Both directory and file exist
-    assert!(logger.set_log_path("test_log_directory/other.log"));
+    assert!(logger.set_log_path("test_log_directory/other.log").is_ok());
     logger.info("Test of log file creation ran successfully, if this is showing up in test.log then everything is working smoothly");
     // note it is correct that this assertion is after the info because there
     // is no guarantee that the writer will be created before the info is logged